[Feat] SectionOfWeatherData enum 위치 WeatherViewModel.swift 내로 이동
